rename bitcoin_core_sv2 common to runtime_api
