+++ bitcoin-core-sv2/README.md
@@ -10,7 +10,7 @@ A Rust library that integrates [Bitcoin Core](https://bitcoin.org/en/bitcoin-cor
 
 The crate exposes three module families:
 
-- `bitcoin_core_sv2::common` - version-agnostic enum-dispatch runtime handles and protocol-specific `new(version, ...)` factories.
+- `bitcoin_core_sv2::runtime_api` - version-agnostic enum-dispatch runtime handles and protocol-specific `new(version, ...)` factories.
 - `bitcoin_core_sv2::unix_capnp::v30x` - Bitcoin Core v30.x IPC implementation.
 - `bitcoin_core_sv2::unix_capnp::v31x` - Bitcoin Core v31.x IPC implementation.
 
@@ -25,7 +25,7 @@ example:
 - `bitcoin_core_sv2::tcp_capnp` (theoretical/future)
 - `bitcoin_core_sv2::http_json_rpc` (theoretical/future)
 
-Downstream applications should integrate through `bitcoin_core_sv2::common`, choose the Bitcoin Core major version at runtime, and build runtimes via `template_distribution_protocol::new` / `job_declaration_protocol::new`.
+Downstream applications should integrate through `bitcoin_core_sv2::runtime_api`, choose the Bitcoin Core major version at runtime, and build runtimes via `template_distribution_protocol::new` / `job_declaration_protocol::new`.
 
 ## Requirements
