+++ bitcoin-core-sv2/src/runtime_api/job_declaration_protocol/io.rs
