+++ bitcoin-core-sv2/src/runtime_api/job_declaration_protocol/mod.rs
@@ -13,7 +13,7 @@
 pub mod io;
 
 use crate::{
-    common::{BitcoinCoreSv2Error, BitcoinCoreSv2Protocol, BitcoinCoreVersion},
+    runtime_api::{BitcoinCoreSv2Error, BitcoinCoreSv2Protocol, BitcoinCoreVersion},
     unix_capnp::{v30x, v31x},
 };
 use async_channel::Receiver;
