+++ bitcoin-core-sv2/src/unix_capnp/v31x/mod.rs
@@ -1,7 +1,7 @@
 //! Bitcoin Core v31.x IPC implementation modules.
 //!
 //! This namespace contains the concrete v31.x runtime implementations used when
-//! [`crate::common::BitcoinCoreVersion::V31X`] is selected.
+//! [`crate::runtime_api::BitcoinCoreVersion::V31X`] is selected.
 //!
 //! It is wired against `bitcoin_capnp_types_v31`, which re-exports the matching `capnp`
 //! and `capnp-rpc` APIs.
