+++ pool-apps/pool/src/lib/downstream/mod.rs
@@ -5,7 +5,7 @@ use std::sync::{
 
 use async_channel::{unbounded, Receiver, Sender};
 use stratum_apps::{
-    bitcoin_core_sv2::common::template_distribution_protocol::CancellationToken,
+    bitcoin_core_sv2::CancellationToken,
     channel_utils::ReceiverCleanup,
     network_helpers::noise_stream::NoiseTcpStream,
     stratum_core::{
