+++ pool-apps/pool/src/lib/template_receiver/sv2_tp/mod.rs
@@ -2,7 +2,7 @@ use std::sync::Arc;
 mod common_message_handler;
 use async_channel::{unbounded, Receiver, Sender};
 use stratum_apps::{
-    bitcoin_core_sv2::common::template_distribution_protocol::CancellationToken,
+    bitcoin_core_sv2::CancellationToken,
     channel_utils::ReceiverCleanup,
     key_utils::Secp256k1PublicKey,
     network_helpers::{self, connect_with_noise, resolve_host_port, TCP_CONNECT_TIMEOUT},
